7463 Create failures package
