+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/nightly/failures/TestFailureException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.drivers.nightly;
+package sleeper.systemtest.drivers.nightly.failures;
 
 public class TestFailureException extends RuntimeException {
